requirements.txt
keyboard
mouse
requests
urllib3
pyautogui
dearpygui
psutil

pyobjc-framework-Quartz
PyGetWindow